+++ src/library/brain/FiringRateModel.h
@@ -44,7 +44,7 @@ class FiringRateModel : public BaseNeuronModel<FiringRateModel__Neuron, FiringRa
 	FiringRateModel( NervousSystem *cns );
 	virtual ~FiringRateModel();
 
-	virtual void init_derived( float initial_activation );
+	virtual void init_derived( double initial_activation );
 
 	virtual void set_neuron( int index,
 							 void *attributes,
