+++ src/library/utils/misc.h
@@ -87,10 +87,10 @@ inline float fmin(float f1, float f2, float f3, float f4, float f5)
 #define clamp(VAL, MIN, MAX) ((VAL) < (MIN) ? (MIN) : ((VAL) > (MAX) ? (MAX) : (VAL)))
 
 
-float logistic(float x, float slope);
-float biasedLogistic(float x, float midpoint, float slope);
-float generalLogistic(float x, float midpoint, float slope, float yneg, float ypos);
-float gaussian( float x, float mean, float variance );
+double logistic(double x, double slope);
+double biasedLogistic(double x, double midpoint, double slope);
+double generalLogistic(double x, double midpoint, double slope, double yneg, double ypos);
+double gaussian( double x, double mean, double variance );
 
 inline float dist( float x1, float y1, float x2, float y2 ) { return sqrt( (x1-x2)*(x1-x2) + (y1-y2)*(y1-y2) ); }
